fix: address PR #369 review — cache module, eviction, CTM-aware dedup
Three follow-ups from Copilot's review on PR #369:

1. BoundedEntryCache eviction-on-replace inefficiency

   Previously, `insert` evicted the oldest entry even when updating an
   existing key. Now checks `Entry::Occupied` first and short-circuits
   with a value replacement — no eviction, preserving cache contents
   and improving hit rate under repeated updates. Also upgraded `get`
   to move accessed entries to the back of the eviction queue (LRU
   semantics instead of pure FIFO).

2. Cross-module coupling — extract cache utilities to `src/cache.rs`

   `BoundedEntryCache` and `MutexExt` were defined in `document.rs`
   but imported by `fonts::cmap`. Moved both to a new `src/cache.rs`
   module. `document.rs` re-exports them via `pub(crate) use` so
   existing internal consumers are unaffected. `fonts::cmap` now
   imports from `crate::cache` directly.

3. CTM-aware XObject dedup in path extraction

   `processed_xobjects` was `HashSet<ObjectRef>` — keyed solely by
   object ID. When the same Form XObject was invoked at multiple
   positions (different CTMs), only the first invocation was processed
   and the rest were silently skipped, losing content (e.g. repeated
   logos, chart legends at different positions).

   Changed the key to `(ObjectRef, [i32; 6])` where the `[i32; 6]`
   is a CTM fingerprint (scale/rotation at 0.01 precision, translation
   at 0.1 precision). Same XObject at the same position is still
   deduplicated (the OOM fix), but same XObject at different positions
   now processes each invocation separately.

Full suite: 6300 passed, 0 failed, 228 ignored.
